Extract readysystem from scorecard
