+++ Assets/00_Content/Scripts/Menu/ReadySystem.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 4816c70c88314a18b2d0c2df676bf6dd
+timeCreated: 1620800058
